// plugin-or-cli/src/types/analytics.ts
export type CommitAnalytics = {
  /**
   * Total number of commits made across all PRs associated to the
   * author of the PR that closed the issue (may have re-done the work etc)
   */
  totalCommits: number;
  /**
   * Total number of lines added across all commits
   */
  linesAdded: number;
  /**
   * Total number of lines removed across all commits
   */
  linesRemoved: number;
  /**
   * Number of commits made by each contributor if multiple contributors
   */
  commitsPerContributor: {
    [login: string]: number;
  };
  /**
   * Time between the first commit of the PR and the last commit of the PR (tracking contributor efficiency)
   */
  timeFromFirstCommitToLastCommit: number;
  /**
   * Time between the last commit of the PR and the PR being closed (tracking reviewer efficiency)
   */
  timeFromLastCommitToClose: number;
  /**
   * Average time between commits (tracking contributor efficiency)
   */
  averageTimeBetweenCommits: number;
  averageCommitsPerDay: number;
  averageCommitsPerWeek: number;
  averageCommitsPerMonth: number;
  totalCommitsPerDay: {
    [date: string]: number;
  };
  totalCommitsPerWeek: {
    [date: string]: number;
  };
  totalCommitsPerMonth: {
    [date: string]: number;
  };
};